tests/wpt/tests/css/css-conditional/container-queries/container-computed.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Containment Test: Computed values of container</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-name">
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-type">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
setup(() => assert_implements_container_queries());

test_computed_value('container', 'initial', 'none');
test_computed_value('container', 'inherit', 'none');
test_computed_value('container', 'unset', 'none');
test_computed_value('container', 'none / inline-size');
test_computed_value('container', 'none / size');
test_computed_value('container', 'inline-size / inline-size');
test_computed_value('container', 'block-size / size');
test_computed_value('container', 'foo / inline-size');
test_computed_value('container', 'foo /inline-size', 'foo / inline-size');
test_computed_value('container', 'foo/ inline-size', 'foo / inline-size');
test_computed_value('container', 'foo/inline-size', 'foo / inline-size');
test_computed_value('container', 'FoO / size');
test_computed_value('container', 'foo bar / size', 'foo bar / size');
test_computed_value('container', 'foo / normal', 'foo');
</script>
